fix all warnings + fix codestyle

+++ Differentiator/Libs/TaylorDecomposition/TaylorDecomposition.h
@@ -3,11 +3,11 @@
 
 #include "../TreeSetup.h"
 
-Node* TaylorDecomp(Node* function, const char* variable, double var_value, unsigned int dec_order);
+Node* TaylorDecomp(Node* function, const char* variable, double var_value, int dec_order);
 
 Node* GetFunctionValueAtPoint(Node* function, const char* variable, double var_value);
 
-Node* GetOneTaylorDecMember(Node* derivative, int fact, int number_member,
+Node* GetOneTaylorDecMember(Node* derivative, long long fact, int number_member,
                                    const char* variable, double var_value);
 
 #endif //__SYM_TAYLOR__
